Added do query resource tests.
Removed the difference in columns do to RateType.  This simplifies the code a lot

Removed the "txt" version of doQuery because I could not figure out how to tell the restlet Request to request the binary representation

Since the logging for the restlet component is still broken because of mysterious logging bridge issues that I can no longer research from the restlet wiki because it disappeared.  I am expanding the resource tests so that the server functionality is thoroughly tested without having to guess why the component failed