Added trapping of the Process error and output streams
Turns out that MySQL and Derby are totally different in handling the creation and removal of databases.  Got rid of the kludgy DbManagement code and expended the filters to handle db create and remove operations.

Added an extra test testDbPools which creates and removes several different databases. This led to a re-thinking on how the experiment name is injected into the code.  It turns out that the best approach is to have some kind of dummy name and then set the real names to the database specific classes.  This leaves DbPools in an odd spot since it is also injected but is agnostic to the current experiment.

Finally figured out how to manage the network start/stop of derbyDb.  It turns out that the shell script starts its own JVM and you have to use the DERBY_OPTS shell variable to specify the data directory.  I also had to extend the File class to delete a folder and its contents.

Used the @Before/AfterClass to allow derby to be started both for the suite and for the individual tests without leaving crud behind.  The key is to ping derby with a shell script and then keep track of whether derby was started for a particular test class or not.